refactor: change structure modal and fuctional
