+++ package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^2.2.6",
     "@sveltejs/vite-plugin-svelte": "^3.1.1",
+    "file-saver": "^2.0.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-error-boundary": "^4.0.13",
@@ -27,6 +28,7 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^6.4.6",
     "@testing-library/react": "^16.0.0",
+    "@types/file-saver": "^2.0.7",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "@types/react-redux": "^7.1.33",
